separate utils from preview to storybook-utils
