+++ sdk/.storybook/preview.tsx
@@ -6,28 +6,19 @@ import { ThemeProvider } from '@0xsequence/design-system';
 import type { Preview } from '@storybook/react-vite';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
-import {
-	anvil,
-	foundry,
-	mainnet as wagmiMainnet,
-	polygon as wagmiPolygon,
-} from 'viem/chains';
-import { createConfig, http, WagmiProvider } from 'wagmi';
-import { mock } from 'wagmi/connectors';
+import { http, WagmiProvider } from 'wagmi';
 import { MarketplaceProvider } from '../src/react/provider';
 import { ModalProvider } from '../src/react/ui/modals/modal-provider';
 import '../src/index.css';
 import { initialize, mswLoader } from 'msw-storybook-addon';
 import {
-	type Client,
-	createPublicClient,
-	createTestClient,
-	createWalletClient,
-	publicActions,
-	walletActions,
-} from 'viem';
-import * as chains from 'viem/chains';
+	createTestQueryClient,
+	sequenceConnectConfig,
+	wagmiConfig,
+} from '../test/storybook-utils';
 import { ConnectionStatus } from './ConnectionStatus';
+
+const testQueryClient = createTestQueryClient();
 
 // Initialize MSW
 initialize({
@@ -39,124 +30,6 @@ initialize({
 		}
 		// Bypass all other requests (fonts, static assets, etc.)
 	},
-});
-
-// Use the same mock connector pattern as test-utils.tsx
-const mockConnector = mock({
-	accounts: ['0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266'],
-	features: {
-		defaultConnected: true,
-		reconnect: true,
-	},
-});
-
-const mainnet = {
-	...wagmiMainnet,
-	rpcUrls: { default: { http: ['http://127.0.0.1:8545/1'] } },
-};
-
-const polygon = {
-	...wagmiPolygon,
-	rpcUrls: { default: { http: ['http://127.0.0.1:8545/1'] } },
-};
-
-export const TEST_CHAIN = {
-	...anvil,
-	rpcUrls: {
-		default: { http: ['http://127.0.0.1:8545/1'] },
-	},
-};
-
-// Default accounts for Anvil
-export const TEST_ACCOUNTS = [
-	'0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266',
-	'0x70997970C51812dc3A010C7d01b50e0d17dc79C8',
-	'0x3C44CdDdB6a900fa2b585dd299e03d12FA4293BC',
-	'0x90F79bf6EB2c4f870365E785982E1f101E93b906',
-	'0x15d34AAf54267DB7D7c367839AAf71A00a2C6A65',
-	'0x9965507D1a55bcC2695C58ba16FB37d819B0A4dc',
-	'0x976EA74026E726554dB657fA54763abd0C3a0aa9',
-	'0x14dC79964da2C08b23698B3D3cc7Ca32193d9955',
-	'0x23618e81E3f5cdF7f54C3d65f7FBc0aBf5B21E8f',
-	'0xa0Ee7A142d267C1f36714E4a8F75612F20a79720',
-] as const;
-
-export const TEST_PRIVATE_KEYS = [
-	'0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80',
-	'0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d',
-	'0x5de4111afa1a4b94908f83103eb1f1706367c2e68ca870fc3fb9a804cdab365a',
-	'0x7c852118294e51e653712a81e05800f419141751be58f605c371e15141b007a6',
-	'0x47e179ec197488593b187f80a00eb0da91f1b9d0b13f8733639f19c30a34926a',
-	'0x8b3a350cf5c34c9194ca85829a2df0ec3153be0318b5e2d3348e872092edffba',
-	'0x92db14e403b83dfe3df233f83dfa3a0d7096f21ca9b0d6d6b8d88b2b4ec1564e',
-	'0x4bbbf85ce3377467afe5d46f804f221813b2bb87f24d81f60f1fcdbf7cbf4356',
-	'0xdbda1821b80551c9d65939329250298aa3472ba22feea921c0cf5d620ea67b97',
-	'0x2a871d0798f97d79848a013d4936a73bf4cc922c825d33c1cf7073dff6d409c6',
-] as const;
-
-export const getMockWalletClient = () =>
-	createWalletClient({
-		transport: http(foundry.rpcUrls.default.http[0]),
-		chain: foundry,
-		account: '0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266',
-		key: '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80',
-		pollingInterval: 100,
-	});
-
-export const getPublicClient = ({
-	chainId = foundry.id,
-}: {
-	chainId?: number;
-}) => {
-	const chain = Object.entries(chains).find(
-		([_, chain]) => chain.id === chainId,
-	)?.[1];
-	if (!chain) throw new Error(`Chain ${chainId} not found`);
-
-	return createPublicClient({
-		transport: http(foundry.rpcUrls.default.http[0]),
-		chain,
-		pollingInterval: 100,
-	});
-};
-
-const testClient = createTestClient({
-	transport: http(),
-	chain: TEST_CHAIN,
-	mode: 'anvil',
-	account: TEST_ACCOUNTS[0],
-	key: TEST_PRIVATE_KEYS[0],
-	pollingInterval: 100,
-})
-	.extend(publicActions)
-	.extend(walletActions) satisfies Client;
-
-const wagmiConfig = createConfig({
-	chains: [mainnet, polygon],
-	connectors: [mockConnector],
-	transports: {
-		[mainnet.id]: http(),
-		[polygon.id]: http(),
-	},
-	multiInjectedProviderDiscovery: false,
-});
-
-// Use the same QueryClient configuration as test-utils.tsx
-const queryClient = new QueryClient({
-	defaultOptions: {
-		queries: {
-			retry: false,
-			staleTime: 0,
-		},
-	},
-});
-
-// Use the same sequence connect config pattern as test-utils.tsx
-const sequenceConnectConfig = createSequenceConnectConfig('universal', {
-	projectAccessKey: 'test',
-	chainIds: [1, 137],
-	defaultChainId: 1,
-	appName: 'Demo Dapp',
 });
 
 // Mock marketplace SDK config for Storybook
@@ -188,7 +61,7 @@ const preview: Preview = {
 		(Story) => {
 			return (
 				<WagmiProvider config={wagmiConfig}>
-					<QueryClientProvider client={queryClient}>
+					<QueryClientProvider client={testQueryClient}>
 						<SequenceConnectProvider
 							config={sequenceConnectConfig.connectConfig}
 						>
